requirements.txt
justpy
python-dotenv